requirements.txt
requests
dotenv
pandas
gspread
oauth2client
mysql-connector-python
